Check usersession is null

+++ HareDB_HBaseClient_FacadeAPI/src/main/java/com/haredb/harespark/operator/HareSparkOperator.java
@@ -37,6 +37,7 @@ public class HareSparkOperator {
 	private UserSessionBean userSessionBean;
 	private HareSparkFacade hareSparkFacade;
 	
+	public final static String USERSESSIONNULLMSG = "usersession is null, please connect to usersession restful";
 	public HareSparkOperator() {
 	
 	}
@@ -70,9 +71,16 @@ public class HareSparkOperator {
 		
 		startTime = System.currentTimeMillis();
 		try {
-			hareSparkFacade = new HareSparkFacade(userSessionBean.getConfigurationFolderPath(), sysConfig);
-			hareSparkFacade.createTable(createTableBean.getTablename(), createTableBean.getColumnNames(), createTableBean.getDataTypes());
-			bean.setStatus(bean.SUCCESS);
+			
+			if(userSessionBean == null){
+				bean.setStatus(bean.ERROR);
+				bean.setException(USERSESSIONNULLMSG);
+			}else{
+				hareSparkFacade = new HareSparkFacade(userSessionBean.getConfigurationFolderPath(), sysConfig);
+				hareSparkFacade.createTable(createTableBean.getTablename(), createTableBean.getColumnNames(), createTableBean.getDataTypes());
+				bean.setStatus(bean.SUCCESS);
+			}
+			
 		}catch(Exception e) {
 			bean.setStatus(bean.ERROR);
 			bean.setException("Create table "+ createTableBean.getTablename() + " failed : " + e.getMessage());
@@ -87,20 +95,25 @@ public class HareSparkOperator {
 		UploadDataFileResponseBean bean = new UploadDataFileResponseBean();
 		startTime = System.currentTimeMillis();		
 		// Todo : change to thread	
-		try {			
-			hareSparkFacade = new HareSparkFacade(userSessionBean.getConfigurationFolderPath(), sysConfig);
-			boolean skipHeader = false;
-			try {
-				skipHeader = Boolean.valueOf(uploadDataFileBean.getSkipHeader());
-			}catch (Exception e) {
-				throw new Exception("skipHeader value " + uploadDataFileBean.getSkipHeader() + " is error");
-			}			
-			asychronizeRunnable runnable = new asychronizeRunnable(hareSparkFacade, uploadDataFileBean);
-			new Thread(runnable).start();			
-//			hareSparkFacade.uploadDataFile(uploadDataFileBean.getDataFilePath(), uploadDataFileBean.getResultPath(), uploadDataFileBean.getCsvSeparator(), uploadDataFileBean.getTableName(), skipHeader, uploadDataFileBean.getOperator());
-			
-			bean.setUploadJobName(uploadDataFileBean.getResultPath());
-			bean.setStatus(bean.SUCCESS);
+		try {		
+			if(userSessionBean == null){
+				bean.setStatus(bean.ERROR);
+				bean.setException(USERSESSIONNULLMSG);
+			}else{
+				hareSparkFacade = new HareSparkFacade(userSessionBean.getConfigurationFolderPath(), sysConfig);
+				boolean skipHeader = false;
+				try {
+					skipHeader = Boolean.valueOf(uploadDataFileBean.getSkipHeader());
+				}catch (Exception e) {
+					throw new Exception("skipHeader value " + uploadDataFileBean.getSkipHeader() + " is error");
+				}			
+				asychronizeRunnable runnable = new asychronizeRunnable(hareSparkFacade, uploadDataFileBean);
+				new Thread(runnable).start();			
+	//			hareSparkFacade.uploadDataFile(uploadDataFileBean.getDataFilePath(), uploadDataFileBean.getResultPath(), uploadDataFileBean.getCsvSeparator(), uploadDataFileBean.getTableName(), skipHeader, uploadDataFileBean.getOperator());
+				
+				bean.setUploadJobName(uploadDataFileBean.getResultPath());
+				bean.setStatus(bean.SUCCESS);
+			}
 		}catch(Exception e) {
 			bean.setStatus(bean.ERROR);
 			bean.setException("Upload DataFile failed : " + e.getMessage());
@@ -116,14 +129,19 @@ public class HareSparkOperator {
 		UploadDataFileStatusResponseBean bean = new UploadDataFileStatusResponseBean();
 		startTime = System.currentTimeMillis();		
 		try {
-			hareSparkFacade = new HareSparkFacade(userSessionBean.getConfigurationFolderPath(), sysConfig);
-			String response = hareSparkFacade.uploadFileStatus(uploadDataFileStatusBean.getUploadJobName());
-			
-			String[] responseinfo = response.split(",");
-			bean.setStartTime(responseinfo[0]);
-			bean.setFinishTime(responseinfo[1]);
-			bean.setDataStatus(responseinfo[2]);
-			bean.setStatus(bean.SUCCESS);
+			if(userSessionBean == null){
+				bean.setStatus(bean.ERROR);
+				bean.setException(USERSESSIONNULLMSG);
+			}else{
+				hareSparkFacade = new HareSparkFacade(userSessionBean.getConfigurationFolderPath(), sysConfig);
+				String response = hareSparkFacade.uploadFileStatus(uploadDataFileStatusBean.getUploadJobName());
+				
+				String[] responseinfo = response.split(",");
+				bean.setStartTime(responseinfo[0]);
+				bean.setFinishTime(responseinfo[1]);
+				bean.setDataStatus(responseinfo[2]);
+				bean.setStatus(bean.SUCCESS);
+			}
 		}catch(Exception e) {
 			bean.setStatus(bean.ERROR);
 			bean.setException("Query DataFile Status failed : " + e.getMessage());
@@ -140,14 +158,17 @@ public class HareSparkOperator {
 		startTime = System.currentTimeMillis();		
 		try {
 			String jobname = querySubmitBean.getTablename() + "_" + System.currentTimeMillis();
-			
-			
-			hareSparkFacade = new HareSparkFacade(userSessionBean.getConfigurationFolderPath(), sysConfig);
-			asychronizeRunnable runnable = new asychronizeRunnable(hareSparkFacade, querySubmitBean, jobname);
-			new Thread(runnable).start();
-//			hareSparkFacade.queryHDFSTable(querySubmitBean.getResultFileFolder(), querySubmitBean.getSql(), querySubmitBean.getTablename(), jobname);
-			bean.setQueryJobName(jobname);
-			bean.setStatus(bean.SUCCESS);
+			if(userSessionBean == null){
+				bean.setStatus(bean.ERROR);
+				bean.setException(USERSESSIONNULLMSG);
+			}else{
+				hareSparkFacade = new HareSparkFacade(userSessionBean.getConfigurationFolderPath(), sysConfig);
+				asychronizeRunnable runnable = new asychronizeRunnable(hareSparkFacade, querySubmitBean, jobname);
+				new Thread(runnable).start();
+	//			hareSparkFacade.queryHDFSTable(querySubmitBean.getResultFileFolder(), querySubmitBean.getSql(), querySubmitBean.getTablename(), jobname);
+				bean.setQueryJobName(jobname);
+				bean.setStatus(bean.SUCCESS);
+			}
 		}catch(Exception e) {
 			bean.setStatus(bean.ERROR);
 			bean.setException("QuerySubmit failed : " + e.getMessage());
@@ -164,16 +185,20 @@ public class HareSparkOperator {
 		QueryStatusResponseBean bean = new QueryStatusResponseBean();
 		startTime = System.currentTimeMillis();		
 		try {
-			hareSparkFacade = new HareSparkFacade(userSessionBean.getConfigurationFolderPath(), sysConfig);			
-			List<String> responselist = hareSparkFacade.queryHDFSTableStatus(queryStatusBean.getQueryJobName());
-			bean.setJobStatus(responselist.get(0));
-			bean.setJobID(responselist.get(1));
-			bean.setJobName(responselist.get(2));
-			bean.setJobFinishTime(responselist.get(3));
-			bean.setJobStartTime(responselist.get(4));
-			bean.setJobProgress(responselist.get(5));
-			
-			bean.setStatus(bean.SUCCESS);
+			if(userSessionBean == null){
+				bean.setStatus(bean.ERROR);
+				bean.setException(USERSESSIONNULLMSG);
+			}else{
+				hareSparkFacade = new HareSparkFacade(userSessionBean.getConfigurationFolderPath(), sysConfig);			
+				List<String> responselist = hareSparkFacade.queryHDFSTableStatus(queryStatusBean.getQueryJobName());
+				bean.setJobStatus(responselist.get(0));
+				bean.setJobID(responselist.get(1));
+				bean.setJobName(responselist.get(2));
+				bean.setJobFinishTime(responselist.get(3));
+				bean.setJobStartTime(responselist.get(4));
+				bean.setJobProgress(responselist.get(5));
+				bean.setStatus(bean.SUCCESS);
+			}
 		}catch(Exception e) {
 			bean.setStatus(bean.ERROR);
 			bean.setException("QueryStatus failed : " + e.getMessage());
@@ -190,38 +215,41 @@ public class HareSparkOperator {
 		PreviewResponseBean bean = new PreviewResponseBean();
 		startTime = System.currentTimeMillis();		
 		try {
-			hareSparkFacade = new HareSparkFacade(userSessionBean.getConfigurationFolderPath(), sysConfig);
-			int pagesize = 0;
-			int limit = 0;
-			
-			try {
-				pagesize = Integer.valueOf(previewBean.getPageSize());
-				limit = Integer.valueOf(previewBean.getLimit());
-			}catch (Exception e) {
-				throw new Exception(e.getMessage());
-			}			
-			List<String> responselist = hareSparkFacade.previewData(previewBean.getTablename(), pagesize, limit);				
-			if (responselist == null || responselist.size() == 0) {
-				throw new Exception("table " + previewBean.getTablename() + " is not created");
-			}		
-			
-			String[] header = responselist.get(0).split(",");
-			
-			String[][] results = new String[responselist.size() - 1][];
-			if (responselist.size() != 1) {
-				for(int i = 1; i < responselist.size(); i++) {
-					String[] temp = responselist.get(i).split(",");
-					results[i - 1] = new String[temp.length];
-					for (int j = 0; j < temp.length; j++) {
-						results[i - 1][j] = temp[j];
-					}					
+			if(userSessionBean == null){
+				bean.setStatus(bean.ERROR);
+				bean.setException(USERSESSIONNULLMSG);
+			}else{
+				hareSparkFacade = new HareSparkFacade(userSessionBean.getConfigurationFolderPath(), sysConfig);
+				int pagesize = 0;
+				int limit = 0;
+				
+				try {
+					pagesize = Integer.valueOf(previewBean.getPageSize());
+					limit = Integer.valueOf(previewBean.getLimit());
+				}catch (Exception e) {
+					throw new Exception(e.getMessage());
+				}			
+				List<String> responselist = hareSparkFacade.previewData(previewBean.getTablename(), pagesize, limit);				
+				if (responselist == null || responselist.size() == 0) {
+					throw new Exception("table " + previewBean.getTablename() + " is not created");
+				}		
+				
+				String[] header = responselist.get(0).split(",");
+				
+				String[][] results = new String[responselist.size() - 1][];
+				if (responselist.size() != 1) {
+					for(int i = 1; i < responselist.size(); i++) {
+						String[] temp = responselist.get(i).split(",");
+						results[i - 1] = new String[temp.length];
+						for (int j = 0; j < temp.length; j++) {
+							results[i - 1][j] = temp[j];
+						}					
+					}
 				}
+				bean.setHeads(header);
+				bean.setResults(results);
+				bean.setStatus(bean.SUCCESS);
 			}
-			
-			
-			bean.setHeads(header);
-			bean.setResults(results);
-			bean.setStatus(bean.SUCCESS);
 		}catch(Exception e) {
 			bean.setStatus(bean.ERROR);
 			bean.setException("QuerySubmit failed : " + e.getMessage());
@@ -238,9 +266,15 @@ public class HareSparkOperator {
 		startTime = System.currentTimeMillis();		
 
 		try {
-			hareSparkFacade = new HareSparkFacade(userSessionBean.getConfigurationFolderPath(), sysConfig);
-			boolean result = hareSparkFacade.deleteDataFile(deleteDataFileBean.getTablename(), deleteDataFileBean.getDeleteDataFileName());			
-			bean.setStatus(bean.SUCCESS);			
+			if(userSessionBean == null){
+				bean.setStatus(bean.ERROR);
+				bean.setException(USERSESSIONNULLMSG);
+			}else{
+				hareSparkFacade = new HareSparkFacade(userSessionBean.getConfigurationFolderPath(), sysConfig);
+				boolean result = hareSparkFacade.deleteDataFile(deleteDataFileBean.getTablename(), deleteDataFileBean.getDeleteDataFileName());			
+				bean.setStatus(bean.SUCCESS);
+			}
+			
 		}catch(Exception e) {
 			bean.setStatus(bean.ERROR);
 			bean.setException("Delete DataFile failed : " + e.getMessage());
@@ -257,9 +291,14 @@ public class HareSparkOperator {
 		startTime = System.currentTimeMillis();		
 
 		try {
-			hareSparkFacade = new HareSparkFacade(userSessionBean.getConfigurationFolderPath(), sysConfig);
-			hareSparkFacade.dropTable(dropTableBean.getTablename());
-			bean.setStatus(bean.SUCCESS);
+			if(userSessionBean == null){
+				bean.setStatus(bean.ERROR);
+				bean.setException(USERSESSIONNULLMSG);
+			}else{
+				hareSparkFacade = new HareSparkFacade(userSessionBean.getConfigurationFolderPath(), sysConfig);
+				hareSparkFacade.dropTable(dropTableBean.getTablename());
+				bean.setStatus(bean.SUCCESS);
+			}
 		}catch(Exception e) {
 			bean.setStatus(bean.ERROR);
 			bean.setException("Drop table " + dropTableBean.getTablename() + " failed : " + e.getMessage());
@@ -275,9 +314,14 @@ public class HareSparkOperator {
 		ResponseInfoBean bean = new ResponseInfoBean();
 		startTime = System.currentTimeMillis();		
 		try {
-			hareSparkFacade = new HareSparkFacade(userSessionBean.getConfigurationFolderPath(), sysConfig);
-			hareSparkFacade.alterTable(alterTableBean.getTablename(), alterTableBean.getColumnNames(), alterTableBean.getDataTypes());
-			bean.setStatus(bean.SUCCESS);
+			if(userSessionBean == null){
+				bean.setStatus(bean.ERROR);
+				bean.setException(USERSESSIONNULLMSG);	
+			}else{
+				hareSparkFacade = new HareSparkFacade(userSessionBean.getConfigurationFolderPath(), sysConfig);
+				hareSparkFacade.alterTable(alterTableBean.getTablename(), alterTableBean.getColumnNames(), alterTableBean.getDataTypes());
+				bean.setStatus(bean.SUCCESS);
+			}
 		}catch(Exception e) {
 			bean.setStatus(bean.ERROR);
 			bean.setException("Alter table " + alterTableBean.getTablename() +" failed : " + e.getMessage());
@@ -293,23 +337,28 @@ public class HareSparkOperator {
 		DescribeTableResponseBean bean = new DescribeTableResponseBean();
 		startTime = System.currentTimeMillis();		
 		try {
-			hareSparkFacade = new HareSparkFacade(userSessionBean.getConfigurationFolderPath(), sysConfig);
-			TableInfo tableinfo = hareSparkFacade.describeTable(describeTableBean.getTablename());			
-			
-			if (tableinfo == null) {
-				throw new Exception("Table " + describeTableBean.getTablename() + " is not existed");
+			if(userSessionBean == null){
+				bean.setStatus(bean.ERROR);
+				bean.setException(USERSESSIONNULLMSG);
+			}else{
+				hareSparkFacade = new HareSparkFacade(userSessionBean.getConfigurationFolderPath(), sysConfig);
+				TableInfo tableinfo = hareSparkFacade.describeTable(describeTableBean.getTablename());			
+				
+				if (tableinfo == null) {
+					throw new Exception("Table " + describeTableBean.getTablename() + " is not existed");
+				}
+				List<String> listColumns = new ArrayList<String>();
+				List<String> listDataType = new ArrayList<String>();
+				List<TableColumn> list = tableinfo.getColumns();
+				for (TableColumn col : list) {
+					listColumns.add(col.getColumnName());
+					listDataType.add(col.getDataType());
+				}
+				
+				bean.setColumnNames(listColumns);
+				bean.setDataTypes(listDataType);
+				bean.setStatus(bean.SUCCESS);
 			}
-			List<String> listColumns = new ArrayList<String>();
-			List<String> listDataType = new ArrayList<String>();
-			List<TableColumn> list = tableinfo.getColumns();
-			for (TableColumn col : list) {
-				listColumns.add(col.getColumnName());
-				listDataType.add(col.getDataType());
-			}
-			
-			bean.setColumnNames(listColumns);
-			bean.setDataTypes(listDataType);
-			bean.setStatus(bean.SUCCESS);
 		}catch(Exception e) {
 			bean.setStatus(bean.ERROR);
 			bean.setException("Describe table " + describeTableBean.getTablename() + " failed : " + e.getMessage());
@@ -324,6 +373,7 @@ public class HareSparkOperator {
 	public boolean isExists(String tableName){
 		hareSparkFacade = new HareSparkFacade(userSessionBean.getConfigurationFolderPath(), sysConfig);
 		return hareSparkFacade.isExists(tableName);
+		
 	}
 	
 	private class asychronizeRunnable implements Runnable {
